Pull request aims to reduce warnings
The distribution has circular references and global variables. These
throw warnings on compile checks. This is a first step to untangling
those.

It prevents a number of modules to litter the namespace by changing
the use into require: this incurs a penalty at runtime, but because
they are called circular, they will be imported into the namespace
very quickly.

The second is the introduction of a l4p::Global module where global
variables (yuk) reside. Because these variables were littered all
over the place, they caused warnings of themselves. The l4p::Global
does not reference any modules, as not to cause new circ refs.